move function and proposal types to transaction module
