+++ executor/src/wasm_executor.rs
@@ -294,9 +294,8 @@ mod tests {
 	use native_runtime::runtime::staking::balance;
 	use state_machine::TestExternalities;
 	use primitives::{twox_128, AccountId};
-	use primitives::runtime_function::Function;
 	use primitives::block::Header;
-	use primitives::transaction::{Transaction, UncheckedTransaction};
+	use primitives::transaction::{Transaction, UncheckedTransaction, Function};
 	use runtime_std;
 	use ed25519::Pair;
